+++ Classes/Controller/TeaController.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace TTN\Tea\Controller;
 
+use Psr\Http\Message\ResponseInterface;
 use TTN\Tea\Domain\Model\Product\Tea;
 use TTN\Tea\Domain\Repository\Product\TeaRepository;
 use TYPO3\CMS\Extbase\Mvc\Controller\ActionController;
@@ -13,23 +14,22 @@ use TYPO3\CMS\Extbase\Mvc\Controller\ActionController;
  */
 class TeaController extends ActionController
 {
-    /**
-     * @var TeaRepository
-     */
-    private $teaRepository;
+    private TeaRepository $teaRepository;
 
     public function injectTeaRepository(TeaRepository $teaRepository): void
     {
         $this->teaRepository = $teaRepository;
     }
 
-    public function indexAction(): void
+    public function indexAction(): ResponseInterface
     {
         $this->view->assign('teas', $this->teaRepository->findAll());
+        return $this->htmlResponse();
     }
 
-    public function showAction(Tea $tea): void
+    public function showAction(Tea $tea): ResponseInterface
     {
         $this->view->assign('tea', $tea);
+        return $this->htmlResponse();
     }
 }